refactor: scope and demand and custom analytics folder structure
