+++ apps/app/components/analytics/custom-analytics/table.tsx
