+++ apps/app/components/analytics/index.ts
@@ -1,5 +1,3 @@
-export * from "./graph";
-export * from "./create-update-analytics-modal";
+export * from "./custom-analytics";
+export * from "./scope-and-demand";
 export * from "./modal";
-export * from "./sidebar";
-export * from "./table";
