+++ apps/app/components/analytics/scope-and-demand/index.ts
@@ -1,0 +1,3 @@
+export * from "./demand";
+export * from "./scope-and-demand";
+export * from "./scope";
