+++ apps/app/components/analytics/scope-and-demand/scope-and-demand.tsx
@@ -1,0 +1,56 @@
+import { useRouter } from "next/router";
+
+import useSWR from "swr";
+
+// services
+import analyticsService from "services/analytics.service";
+// components
+import { AnalyticsDemand, AnalyticsScope } from "components/analytics";
+// ui
+import { PrimaryButton } from "components/ui";
+// fetch-keys
+import { DEFAULT_ANALYTICS } from "constants/fetch-keys";
+
+export const ScopeAndDemand: React.FC = () => {
+  const router = useRouter();
+  const { workspaceSlug } = router.query;
+
+  const {
+    data: defaultAnalytics,
+    error: defaultAnalyticsError,
+    mutate: mutateDefaultAnalytics,
+  } = useSWR(
+    workspaceSlug ? DEFAULT_ANALYTICS(workspaceSlug.toString()) : null,
+    workspaceSlug ? () => analyticsService.getDefaultAnalytics(workspaceSlug.toString()) : null
+  );
+
+  return (
+    <>
+      {!defaultAnalyticsError ? (
+        defaultAnalytics ? (
+          <div className="h-full overflow-y-auto p-5 text-sm">
+            <div className="grid grid-cols-1 gap-5 lg:grid-cols-2">
+              <AnalyticsDemand defaultAnalytics={defaultAnalytics} />
+              <AnalyticsScope defaultAnalytics={defaultAnalytics} />
+            </div>
+          </div>
+        ) : (
+          <div className="grid h-full place-items-center p-5">
+            <div className="space-y-4 text-brand-secondary">
+              <p className="text-sm">Loading analytics...</p>
+            </div>
+          </div>
+        )
+      ) : (
+        <div className="grid h-full place-items-center p-5">
+          <div className="space-y-4 text-brand-secondary">
+            <p className="text-sm">There was some error in fetching the data.</p>
+            <div className="flex items-center justify-center gap-2">
+              <PrimaryButton onClick={mutateDefaultAnalytics}>Refresh</PrimaryButton>
+            </div>
+          </div>
+        </div>
+      )}
+    </>
+  );
+};
