+++ apps/app/constants/fetch-keys.ts
@@ -173,5 +173,5 @@ export const ANALYTICS = (workspaceSlug: string, params: IAnalyticsParams) =>
   `ANALYTICS${workspaceSlug.toUpperCase()}_${params.x_axis}_${params.y_axis}_${params.segment}_${
     params.project
   }`;
-export const DEFAULT_ANALYTICS = (workspaceSlug: string, params: IAnalyticsParams) =>
+export const DEFAULT_ANALYTICS = (workspaceSlug: string, params?: IAnalyticsParams) =>
   `DEFAULT_ANALYTICS_${workspaceSlug.toUpperCase()}`;
